Changes to be committed: 	modified:   _legacy_docker/how-to-guides/deployment/building-your-service.md 	modified:   _legacy_docker/how-to-guides/deployment/image-repository.md 	modified:   _maestro/how-to-guides/deployment/image-repository.md 	modified:   _skycap/how-to-guides/building/building-your-service.md
Removed the help page for gcr.io docker image repository from our help
pages, as we don't support them anymore due to a change on Gcloud. there
is a ticket created for that in Central issues so that it will be
implemented later. For the time being to reduce the workload on our side
I've removed it.